Add files via upload
The goal of this project is to predict, from the images of a person’s face present in dataset, which emotion the facial expression conveys. Our evaluation metric will be the accuracy for each emotion, supplemented by a confusion matrix which highlights which emotions are better recognized than others. 

Input : Images of 7 types of facial expressions including happy, sad, surprise, fear, anger, disgust, and neutral.
Output : Emotion conveyed by facial expression in image.